# frontend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Powered Multi-Agent Career Opportunity Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/applications": {
      "get": {
        "tags": [
          "tracking"
        ],
        "summary": "List Applications",
        "operationId": "list_applications_applications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApplicationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationResponse"
                  },
                  "title": "Response List Applications Applications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/applications/{application_id}": {
      "get": {
        "tags": [
          "tracking"
        ],
        "summary": "Get Application",
        "operationId": "get_application_applications__application_id__get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/applications/{application_id}/status": {
      "patch": {
        "tags": [
          "tracking"
        ],
        "summary": "Update Application Status",
        "operationId": "update_application_status_applications__application_id__status_patch",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/applications/{application_id}/history": {
      "get": {
        "tags": [
          "tracking"
        ],
        "summary": "Get Application History",
        "operationId": "get_application_history_applications__application_id__history_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationHistoryResponse"
                  },
                  "title": "Response Get Application History Applications  Application Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApplicationHistoryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "from_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          },
          "triggered_by": {
            "type": "string",
            "title": "Triggered By"
          },
          "source_event_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventType"
              },
              {
                "type": "null"
              }
            ]
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "application_id",
          "to_status",
          "changed_at",
          "triggered_by"
        ],
        "title": "ApplicationHistoryResponse"
      },
      "ApplicationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "selected_resume_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Resume Id"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "matched_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Skills",
            "default": []
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills",
            "default": []
          },
          "referral_contact_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Contact Id"
          },
          "applied_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Date"
          },
          "follow_up_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "user_id",
          "company",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ApplicationResponse"
      },
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "DISCOVERED",
          "MATCHED",
          "SHORTLISTED",
          "CONTACT_SEARCH",
          "CONTACT_FOUND",
          "OUTREACH_GENERATED",
          "OUTREACH_APPROVED",
          "OUTREACH_SENT",
          "REFERRED",
          "APPLIED",
          "INTERVIEW",
          "OFFER",
          "REJECTED",
          "IGNORED",
          "WITHDRAWN"
        ],
        "title": "ApplicationStatus",
        "description": "Owner: Tracking Service. See state-machines.md#opportunity-lifecycle-applicationstatus--applicationstatus.\n\nTerminal states: OFFER, REJECTED, IGNORED, WITHDRAWN."
      },
...
      "EventType": {
        "type": "string",
        "enum": [
          "JOB_DISCOVERED",
          "PROFILE_UPDATED",
          "JOB_MATCHED",
          "JOB_SHORTLISTED",
          "CONTACTS_REQUESTED",
          "CONTACTS_FOUND",
          "OUTREACH_GENERATED",
          "OUTREACH_APPROVED",
          "OUTREACH_SENT",
          "APPLICATION_UPDATED"
        ],
        "title": "EventType",
        "description": "Owner: shared, additive only. One member per Kafka event payload\ntype. See event-contracts.md#eventtype-enum-members."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateApplicationStatusRequest": {
        "properties": {
          "new_status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "applied_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "new_status"
        ],
        "title": "UpdateApplicationStatusRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}